This is a project that explores variables.
This project is location based, it required me to interact with the city around in a new way. Simply put, the objective is to measure a variable in San Francisco. You will identify a measurable variable in the city and then create an estimate using the Fermi estimation technique. Next, you will complete the data collection, calculate descriptive statistics on the data, and create relevant data visualizations. You will also have a chance to apply your knowledge of probability and simulation to solve a problem.

This is an individual assignment. Everything you submit should be your own words and reflect your own understanding of the material.